requirements.txt
playwright
beautifulsoup4
pymupdf
ollama
gspread
oauth2client
apscheduler
python-dotenv
pandas
requests
tenacity
pathlib
jinja2
secure-smtplib
fastapi
uvicorn
pydantic
python-multipart
